implemented a first versione of money system and update

+++ Assets/Scripts/SaveObject.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 public class SaveObject
 {
     public float updatedPollution;
-    public float updatedMoney;
+    public long updatedMoney;
     public List<bool> unitsEnabled;
     public List<int> unitsLevel;
 }
